Refactor funcol for readability and dynamo tracing
Move eager kernel impls to separate file, which is eaiser to read
(since users may be confused about 2 versions of each kernel in the same file)
and easier to set a dynamo policy to trace only the first file currently.

[ghstack-poisoned]
